Safari-611.1.10/LayoutTests/scrollingcoordinator/mac/latching/scrolling-select-should-not-latch-mainframe-expected.txt


Tests that scrolling over a <select> doesn't latch on the document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS select.scrollTop is 90
PASS window.pageYOffset is 0
Scrolling up then down in the select
PASS select.scrollTop > 0 is true
PASS window.pageYOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
